+++ hazelcast/src/test/resources/additional-services/com.hazelcast.spi.tenantcontrol.TenantControlFactory
@@ -1,1 +1,1 @@
-com.hazelcast.cache.TenantControlTest$CountingTenantControlFactory
+com.hazelcast.spi.tenantcontrol.TenantControlTestSupport$CountingTenantControlFactory
